refactor: update color handling
