+++ src/webgl/lines.ts
@@ -134,9 +134,10 @@ vec4 getRoundedLineColor(vec4 interiorColor, vec4 borderColor) {
   builder.addFragmentCode(`
 float getLineAlpha() {
   if (uLineEndpointClipRadius > 0.0) {
+    float radiusFactor = 0.99;
     float distFromA = length(vec2(vLineOffsetX * vEdgeLengthInPixels, vLineCoord * vHalfTotalLineWidth));
     float distFromB = length(vec2((1.0 - vLineOffsetX) * vEdgeLengthInPixels, vLineCoord * vHalfTotalLineWidth));
-    if (distFromA < uLineEndpointClipRadius || distFromB < uLineEndpointClipRadius) discard;
+    if (distFromA < uLineEndpointClipRadius * radiusFactor || distFromB < uLineEndpointClipRadius * radiusFactor) discard;
   }
   return clamp((1.0 - abs(vLineCoord)) / vLineFeatherFraction, 0.0, 1.0);
 }
